Move bpdoc.Reader to reader.go
Reader is complicated enough to deserve its own file.
